vgg-face-pytorch/requirements.txt
adversarial_robustness_toolbox==1.11.0
art==5.7
matplotlib==3.2.2
mtcnn==0.1.1
numpy==1.21.6
opencv_python==4.6.0.66
pandas==1.3.5
Pillow==9.2.0
protobuf==3.20.*
torch==1.12.0+cu113
torchfile==0.1.0
torchvision==0.13.0+cu113
